# src/main/resources/swagger/api.yml
# API-first development with OpenAPI
# This file will be used at compile time to generate Spring-MVC endpoint stubs using openapi-generator
openapi: '3.0.1'
info:
  title: 'msUserCreation'
  version: 0.0.1
servers:
  - url: http://localhost:8571/api
    description: Development server
  - url: https://localhost:8571/api
    description: Development server with TLS Profile
paths:
  /user:
    post:
      tags:
        - user
      summary: Add a new user to the store
      operationId: addUser
      requestBody:
        description: User object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        201:
          description: Succesfull Operation, the user was created
          headers:
            Location:
              schema:
                type: string
          content: {}
        303:
          description: Unsuccesfull Operation, the userName already exists
          headers:
            Location:
              schema:
                type: string
          content: {}
        500:
          description: Invalid input
          content: {}
      security:
        - jwt:
            - write:users
            - read:users
      x-codegen-request-body-name: body
  /user/{userId}:
    get:
      tags:
        - user
      summary: Find user by ID
      description: Returns a single User
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: User not found
          content: {}
      security:
        - api_key: []
    put:
      tags:
        - user
      summary: Updates a user in the store with form data
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: User object that needs to be updated to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        405:
          description: Invalid input
          content: {}
      security:
        - jwt:
            - write:users
            - read:users
      x-codegen-request-body-name: body
    delete:
      tags:
        - user
      summary: Deletes a user
      operationId: deleteUser
      parameters:
        - name: api_key
          in: header
          schema:
            type: string
        - name: userId
          in: path
          description: User id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: User not found
          content: {}
      security:
        - jwt:
            - write:users
            - read:users
components:
  responses:
    Problem:
      description: error occurred - see status code and problem object for more information.
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        username:
          maxLength: 20
          minLength: 6
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          description: User Status
          format: int32
      xml:
        name: User
    UserStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        statusCode:
          type: integer
        statusReason:
          type: string

  securitySchemes:
    jwt:
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
security:
  - jwt: []